storage: refactoring to use the new structure
