+++ azurerm/config.go
@@ -16,7 +16,6 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/servicefabric"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/signalr"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/sql"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage"
 )
 
 // ArmClient contains the handles to all the specific Azure Resource Manager
@@ -128,7 +127,6 @@ func (client *ArmClient) Build(o *common.ClientOptions) error {
 	client.Scheduler = scheduler.BuildClient(o)
 	client.SignalR = signalr.BuildClient(o)
 	client.Sql = sql.BuildClient(o)
-	client.Storage = storage.BuildClient(o)
 
 	return nil
 }
