+++ azurerm/internal/provider/services.go
@@ -49,6 +49,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/redis"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/relay"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/resource"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/storage"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/streamanalytics"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/subscription"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/trafficmanager"
@@ -105,6 +106,7 @@ func SupportedServices() []common.ServiceRegistration {
 		relay.Registration{},
 		resource.Registration{},
 
+		storage.Registration{},
 		streamanalytics.Registration{},
 		subscription.Registration{},
 		trafficmanager.Registration{},
